requirements.txt
# Framework Web
Flask==3.1.0
Werkzeug==3.1.3
Jinja2==3.1.4
itsdangerous==2.2.0
blinker==1.9.0

# AI y NLP
openai==1.55.0
tiktoken==0.8.0
pyautogen==0.3.2
FLAML==2.3.2

# Procesamiento de Datos
numpy==1.26.4
pydantic==2.10.0
pydantic_core==2.27.0

# Web Scraping y Requests
beautifulsoup4==4.12.3
requests==2.32.3
httpx==0.27.2
selenium==4.26.1
webdriver-manager==4.0.2

# Utilidades
python-dotenv==1.0.1
tqdm==4.67.0
colorama==0.4.6
termcolor==2.5.0
regex==2024.11.6

# Dependencias del Sistema
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
urllib3==2.2.3
typing_extensions==4.12.2

# Async y Networking
anyio==4.6.2.post1
sniffio==1.3.1
httpcore==1.0.7
h11==0.14.0

# Cache y Storage
diskcache==5.6.3

# Docker Support
docker==7.1.0
websocket-client==1.8.0

# Windows Specific (comentado para Render)
# pywin32==308

# Testing y Debug
attrs==24.2.0
sortedcontainers==2.4.0

# Servidor WSGI
gunicorn==21.2.0

# ... (resto de las dependencias anteriores)